*** empty log message ***

+++ src/passwdentry.cpp
@@ -464,8 +464,8 @@ void passwdentry::print() const {
 
 	printf("Name: %s\n",getName());
 	printf("Password: %s\n",getPassword());
-	printf("User Id: %d\n",getUserId());
-	printf("Primary Group Id: %d\n",getPrimaryGroupId());
+	printf("User Id: %ld\n",(long)getUserId());
+	printf("Primary Group Id: %ld\n",(long)getPrimaryGroupId());
 	printf("Real Name: %s\n",getRealName());
 	printf("Home Directory: %s\n",getHomeDirectory());
 	printf("Shell: %s\n",getShell());
